experiments/mcp-genmedia/sample-agents/adk/requirements.txt
# This file was autogenerated by uv via the following command:
#    uv export --no-hashes --format requirements-txt
annotated-types==0.7.0
    # via pydantic
anyio==4.11.0
    # via
    #   google-genai
    #   httpx
    #   mcp
    #   sse-starlette
    #   starlette
authlib==1.6.5
    # via google-adk
cachetools==6.1.0
    # via google-auth
certifi==2025.8.3
    # via
    #   httpcore
    #   httpx
    #   requests
cffi==1.17.1 ; platform_python_implementation != 'PyPy'
    # via cryptography
charset-normalizer==3.4.3
    # via requests
click==8.3.0
    # via
    #   google-adk
    #   uvicorn
cloudpickle==3.1.1
    # via google-cloud-aiplatform
colorama==0.4.6 ; sys_platform == 'win32'
    # via click
cryptography==45.0.7
    # via authlib
docstring-parser==0.17.0
    # via google-cloud-aiplatform
dotenv==0.9.9
    # via adk-genmedia-agent
fastapi==0.118.3
    # via google-adk
google-adk==1.15.1
    # via adk-genmedia-agent
google-api-core==2.25.2
    # via
    #   google-api-python-client
    #   google-cloud-aiplatform
    #   google-cloud-appengine-logging
    #   google-cloud-bigquery
    #   google-cloud-core
    #   google-cloud-logging
    #   google-cloud-resource-manager
    #   google-cloud-secret-manager
    #   google-cloud-speech
    #   google-cloud-storage
    #   google-cloud-trace
google-api-python-client==2.183.0
    # via google-adk
google-auth==2.40.3
    # via
    #   google-api-core
    #   google-api-python-client
    #   google-auth-httplib2
    #   google-cloud-aiplatform
    #   google-cloud-appengine-logging
    #   google-cloud-bigquery
    #   google-cloud-core
    #   google-cloud-logging
    #   google-cloud-resource-manager
    #   google-cloud-secret-manager
    #   google-cloud-speech
    #   google-cloud-storage
    #   google-cloud-trace
    #   google-genai
google-auth-httplib2==0.2.0
    # via google-api-python-client
google-cloud-aiplatform==1.115.0
    # via google-adk
google-cloud-appengine-logging==1.6.2
    # via google-cloud-logging
google-cloud-audit-log==0.3.2
    # via google-cloud-logging
google-cloud-bigquery==3.38.0
    # via google-cloud-aiplatform
google-cloud-core==2.4.3
    # via
    #   google-cloud-bigquery
    #   google-cloud-logging
    #   google-cloud-storage
google-cloud-logging==3.12.1
    # via google-cloud-aiplatform
google-cloud-resource-manager==1.14.2
    # via google-cloud-aiplatform
google-cloud-secret-manager==2.24.0
    # via google-adk
google-cloud-speech==2.33.0
    # via google-adk
google-cloud-storage==3.2.0
    # via
    #   google-adk
    #   google-cloud-aiplatform
google-cloud-trace==1.16.2
    # via opentelemetry-exporter-gcp-trace
google-crc32c==1.7.1
    # via
    #   google-cloud-storage
    #   google-resumable-media
google-genai==1.29.0
    # via
    #   google-adk
    #   google-cloud-aiplatform
google-resumable-media==2.7.2
    # via
    #   google-cloud-bigquery
    #   google-cloud-storage
googleapis-common-protos==1.70.0
    # via
    #   google-api-core
    #   google-cloud-audit-log
    #   grpc-google-iam-v1
    #   grpcio-status
graphviz==0.21
    # via google-adk
greenlet==3.2.4 ; (python_full_version < '3.14' and platform_machine == 'AMD64') or (python_full_version < '3.14' and platform_machine == 'WIN32') or (python_full_version < '3.14' and platform_machine == 'aarch64') or (python_full_version < '3.14' and platform_machine == 'amd64') or (python_full_version < '3.14' and platform_machine == 'ppc64le') or (python_full_version < '3.14' and platform_machine == 'win32') or (python_full_version < '3.14' and platform_machine == 'x86_64')
    # via sqlalchemy
grpc-google-iam-v1==0.14.2
    # via
    #   google-cloud-logging
    #   google-cloud-resource-manager
    #   google-cloud-secret-manager
grpcio==1.75.1
    # via
    #   google-api-core
    #   googleapis-common-protos
    #   grpc-google-iam-v1
    #   grpcio-status
grpcio-status==1.75.1
    # via google-api-core
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.9
    # via httpx
httplib2==0.31.0
    # via
    #   google-api-python-client
    #   google-auth-httplib2
httpx==0.28.1
    # via
    #   google-genai
    #   mcp
httpx-sse==0.4.3
    # via mcp
idna==3.10
    # via
    #   anyio
    #   httpx
    #   requests
importlib-metadata==8.7.0
    # via opentelemetry-api
mcp==1.23.0
    # via google-adk
numpy==2.3.3
    # via shapely
opentelemetry-api==1.37.0
    # via
    #   google-adk
    #   google-cloud-logging
    #   opentelemetry-exporter-gcp-trace
    #   opentelemetry-resourcedetector-gcp
    #   opentelemetry-sdk
    #   opentelemetry-semantic-conventions
opentelemetry-exporter-gcp-trace==1.9.0
    # via
    #   google-adk
    #   google-cloud-aiplatform
opentelemetry-resourcedetector-gcp==1.9.0a0
    # via opentelemetry-exporter-gcp-trace
opentelemetry-sdk==1.37.0
    # via
    #   google-adk
    #   google-cloud-aiplatform
    #   opentelemetry-exporter-gcp-trace
    #   opentelemetry-resourcedetector-gcp
opentelemetry-semantic-conventions==0.55b1
    # via opentelemetry-sdk
packaging==25.0
    # via
    #   google-cloud-aiplatform
    #   google-cloud-bigquery
proto-plus==1.26.1
    # via
    #   google-api-core
    #   google-cloud-aiplatform
    #   google-cloud-appengine-logging
    #   google-cloud-logging
    #   google-cloud-resource-manager
    #   google-cloud-secret-manager
    #   google-cloud-speech
    #   google-cloud-trace
protobuf==6.31.1
    # via
    #   google-api-core
    #   google-cloud-aiplatform
    #   google-cloud-appengine-logging
    #   google-cloud-audit-log
    #   google-cloud-logging
    #   google-cloud-resource-manager
    #   google-cloud-secret-manager
    #   google-cloud-speech
    #   google-cloud-trace
    #   googleapis-common-protos
    #   grpc-google-iam-v1
    #   grpcio-status
    #   proto-plus
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.2
    # via google-auth
pycparser==2.23 ; platform_python_implementation != 'PyPy'
    # via cffi
pydantic==2.11.10
    # via
    #   fastapi
    #   google-adk
    #   google-cloud-aiplatform
    #   google-genai
    #   mcp
    #   pydantic-settings
pydantic-core==2.39.0
    # via pydantic
pydantic-settings==2.10.1
    # via mcp
pyparsing==3.2.5
    # via httplib2
python-dateutil==2.9.0.post0
    # via google-cloud-bigquery
python-dotenv==1.1.1
    # via
    #   dotenv
    #   google-adk
    #   pydantic-settings
python-multipart==0.0.20
    # via mcp
pyyaml==6.0.3
    # via google-adk
requests==2.32.5
    # via
    #   google-api-core
    #   google-cloud-bigquery
    #   google-cloud-storage
    #   google-genai
    #   opentelemetry-resourcedetector-gcp
rsa==4.9.1
    # via google-auth
shapely==2.1.2
    # via google-cloud-aiplatform
six==1.17.0
    # via python-dateutil
sniffio==1.3.1
    # via anyio
sqlalchemy==2.0.43
    # via google-adk
sse-starlette==3.0.2
    # via mcp
starlette==0.49.1
    # via
    #   fastapi
    #   mcp
    #   sse-starlette
typing-extensions==4.15.0
    # via
    #   fastapi
    #   google-adk
    #   google-cloud-aiplatform
    #   google-genai
    #   opentelemetry-api
    #   opentelemetry-resourcedetector-gcp
    #   opentelemetry-sdk
    #   opentelemetry-semantic-conventions
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
    #   typing-inspection
typing-inspection==0.4.2
    # via
    #   pydantic
    #   pydantic-settings
tzdata==2025.2 ; sys_platform == 'win32'
    # via tzlocal
tzlocal==5.3.1
    # via google-adk
uritemplate==4.2.0
    # via google-api-python-client
urllib3==2.6.0
    # via requests
uvicorn==0.36.1
    # via
    #   google-adk
    #   mcp
websockets==15.0.1
    # via google-genai
wrapt==1.17.3
    # via deprecated
zipp==3.23.0
    # via importlib-metadata
